+++ src/app/components/Input/Input.tsx
@@ -1,14 +1,14 @@
-import { Input } from '@/types/Input';
+import { Input } from '@/types/Signup';
 
 export default function Input({ name, type, placeholder, required }: Input) {
     return (
-        <div className="flex flex-col gap-2 max-w-full font-primary">
+        <div className="flex flex-col gap-2 w-full font-primary font-normal text-lg text-gray-800">
             <label htmlFor={name}>{placeholder}</label>
             <input
                 id={name}
                 name={name}
                 type={type}
-                className="w-full h-16 px-2 bg-white text-black border border-gray-300 rounded-md"
+                className="h-16 px-2 bg-white border border-gray-500 rounded-md"
                 required={required}
             />
         </div>
