--- src/types/Input.ts
@@ -1,8 +1,0 @@
-type Input = {
-    name: string;
-    type: string;
-    placeholder: string;
-    required?: boolean;
-};
-
-export type { Input };
