+++ src/types/Signup.ts
@@ -1,0 +1,14 @@
+type Input = {
+    name: string;
+    type: string;
+    placeholder: string;
+    required?: boolean;
+};
+
+type Button = {
+    title: string;
+    type: any;
+    className: string;
+};
+
+export type { Input, Button };
